The view on the profile screen has been replaced with Jetpack compose

+++ app/src/main/java/com/example/tanorami/builds_hero_from_users/data/model/BuildHero.kt
@@ -1,7 +1,7 @@
 package com.example.tanorami.builds_hero_from_users.data.model
 
 data class BuildHero(
-    val idBuild: Int,
+    val idBuild: Long,
     val idHero: Int,
     val idWeapon: Int,
     val idRelicTwoParts: Int,
